+++ Source/Channel/DtRxBackend.h
@@ -40,7 +40,7 @@ typedef struct DtRxPort
     DtDevice* Device;
     int Port; // From 1
     int PortIndex;
-    uint32_t Caps; // DT_CAP_ flags of the port
+    uint64_t Caps; // DT_CAP_ flags of the port
 } DtRxPort;
 
 // What every side has: its functions, its port, and the receive mode and control, which
